+++ packages/web-router/src/types.ts
@@ -1,3 +1,6 @@
+/**
+ * @fileoverview TypeScript type definitions for web-router
+ */
 import type {
   ActionModule,
   FetchEventLike,
